+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes qsf-integration project will be documented in this file.
 ### Added
 - JsonBuilder: implement a logic to remove json nodes with the json substitutor
 - JsonBuilder: allow to replace multiple json keys
-- PipelineBuilder: allow to set an id for loop and condition filter for better debugging 
+- PipelineBuilder: allow to set an id for loop and condition filter for better debugging
+- SuggestQueryTransoformer: implement qsfql filter
 
 ### Removed